+++ src/cs/implementations/convenience.rs
@@ -8,7 +8,7 @@ use super::verifier::VerificationKey;
 use super::witness::WitnessVec;
 use crate::cs::cs_builder_verifier::CsVerifierBuilder;
 use crate::cs::oracle::merkle_tree::MerkleTreeWithCap;
-use crate::dag::{ResolverSortingMode, CircuitResolver};
+use crate::dag::{CircuitResolver};
 use crate::cs::traits::GoodAllocator;
 
 use super::transcript::Transcript;
