+++ src/dag/mod.rs
@@ -1,28 +1,20 @@
-use std::cell::UnsafeCell;
+
+use self::resolvers::mt::sorters::sorter_runtime::RuntimeResolverSorter;
 use std::fmt::Debug;
 use std::hint::spin_loop;
-use std::rc::Rc;
+use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Mutex};
-use std::thread::JoinHandle;
 
-use bincode::Config;
-
-use crate::config::{CSResolverConfig, DoPerformRuntimeAsserts, Resolver};
+use crate::config::CSResolverConfig;
+use crate::cs::Place;
 use crate::cs::traits::cs::{DstBuffer, CSWitnessSource};
 use crate::field::SmallField;
 
 mod awaiters;
 mod guide;
-mod registrar;
-pub(crate) mod resolution_window;
-pub(crate) mod resolver;
-mod resolver_box;
-pub mod sorter_runtime;
-pub mod sorter_playback;
-pub mod resolvers;
 mod primitives;
-
+mod resolver_box;
+pub mod resolvers;
 
 pub trait TrivialWitnessCastable<F: SmallField, const N: usize>:
     'static + Clone + std::fmt::Debug + Send + Sync
@@ -86,13 +78,6 @@ impl<F: SmallField, const N: usize, S: WitnessSource<F>> CSWitnessValues<F, N, S
     }
 }
 
-use crate::cs::Place;
-use crate::utils::PipeOp;
-
-use self::guide::{GuideOrder, OrderInfo, RegistrationNum};
-use self::resolver::{ResolverIx, ResolverCommonData};
-use self::resolvers::mt::ResolverComms;
-use self::sorter_runtime::RuntimeResolverSorter;
 // we use Arc and interior mutability, so we want Send + Sync just in case
 
 pub trait WitnessSource<F: SmallField>: 'static + Send + Sync {
@@ -110,21 +95,6 @@ pub trait WitnessSourceAwaitable<F: SmallField>: WitnessSource<F> {
 
 pub trait Awaiter<'a> {
     fn wait(&self);
-}
-
-// TODO: delete
-pub trait ResolutionRecordStorage {
-    type Id;
-    fn store(&mut self, id: Self::Id, record: &ResolutionRecord);
-    fn get(&self, id: Self::Id) -> Rc<ResolutionRecord>;
-}
-
-pub trait ResolutionRecordWriter {
-    fn store(&mut self, record: &ResolutionRecord);
-}
-
-pub trait ResolutionRecordSource {
-    fn get(&self) -> &ResolutionRecord;
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -153,98 +123,7 @@ impl From<usize> for CircuitResolverOpts {
     }
 }
 
-pub struct TestRecordStorage {
-    record: std::rc::Rc<ResolutionRecord>
-}
-
-impl ResolutionRecordStorage for TestRecordStorage {
-    type Id = ();
-
-    fn store(&mut self, id: Self::Id, record: &ResolutionRecord) {
-    }
-
-    fn get(&self, id: Self::Id) -> std::rc::Rc<ResolutionRecord> {
-        self.record.clone()
-    }
-}
-
-impl ResolutionRecordSource for TestRecordStorage {
-    fn get(&self) -> &ResolutionRecord {
-        &self.record
-    }
-}
-
-pub struct NullRecordWriter();
-impl ResolutionRecordWriter for NullRecordWriter {
-    fn store(&mut self, record: &ResolutionRecord) {
-    }
-}
-
-#[derive(Default, Clone, Debug)]
-pub struct ResolutionRecordItem { 
-    added_at: RegistrationNum,
-    accepted_at: RegistrationNum,
-    /// The size of the order list when this registration was processed.
-    order_len: usize,
-    order_ix: resolver::OrderIx,
-    parallelism: u16,
-}
-
-#[derive(Clone, Debug)]
-pub struct ResolutionRecord {
-    pub items: Vec<ResolutionRecordItem>,
-    pub registrations_count: usize,
-    pub values_count: usize
-}
-
-impl ResolutionRecord {
-    fn new(registrations_count: usize, values_count: usize, size: usize) -> Self {
-        Self {
-            registrations_count,
-            values_count,
-            items:
-                Vec::with_capacity(size)
-                .op(|x| x.resize_with(size, ResolutionRecordItem::default))
-        }
-    }
-}
-
 pub trait TrackId: From<u64> + Into<u64> + Into<usize> + Eq + Ord + Debug + Default + Clone + Copy {}
-
-pub trait ResolverSortingMode<F: SmallField>: Sized
-{
-    type Arg;
-    type Config: resolution_window::RWConfig<Self::TrackId> + 'static;
-    type TrackId: TrackId + 'static;
-    
-
-    fn new(opts: Self::Arg, comms: Arc<ResolverComms>, debug_track: &Vec<Place>) -> (Self, Arc<ResolverCommonData<F, Self::TrackId>>);
-    fn set_value(&mut self, key: Place, value: F);
-    fn add_resolution<Fn>(&mut self, inputs: &[Place], outputs: &[Place], f: Fn)
-    where
-        Fn: FnOnce(&[F], &mut DstBuffer<'_, '_, F>) + Send + Sync;
-
-    unsafe fn internalize(
-        &mut self, 
-        resolver_ix: ResolverIx,
-        inputs: &[Place], 
-        outputs: &[Place],
-        added_at: RegistrationNum);
-    fn internalize_one(
-        &mut self,
-        resolver_ix: ResolverIx,
-        inputs: &[Place],
-        outputs: &[Place],
-        added_at: RegistrationNum
-    ) -> Vec<ResolverIx>;
-
-    fn flush(&mut self);
-    fn final_flush(&mut self);
-    fn write_sequence(&mut self);
-
-    fn retrieve_sequence(&mut self) -> &ResolutionRecord;
-}
-
 
 pub trait CircuitResolver<
     F: SmallField,
