+++ src/dag/resolvers/mod.rs
@@ -1,8 +1,10 @@
 mod null;
 mod st;
-pub(crate) mod mt;
+pub mod mt;
 
 pub(crate) use null::NullCircuitResolver;
 pub(crate) use mt::MtCircuitResolver;
 pub(crate) use st::StCircuitResolver;
 pub use st::StCircuitResolverParams;
+
+pub use mt::sorters::ResolverSortingMode;
