+++ src/dag/resolvers/mt/mod.rs
@@ -1,145 +1,57 @@
-// Interferes with paranioa mode.
-#![allow(clippy::overly_complex_bool_expr)]
-#![allow(clippy::nonminimal_bool)]
+mod resolution_window;
+pub mod sorters;
+mod registrar;
 
-use super::primitives::Values;
-use super::{ResolverSortingMode, ResolutionRecord, TrackId, CircuitResolver, CircuitResolverOpts};
-use crate::log;
-use std::any::Any;
-use std::cell::{Cell, UnsafeCell};
-use std::fmt::{Display, Debug};
-use std::marker::PhantomData;
+use std::{sync::{Arc, atomic::{AtomicIsize, AtomicBool, fence}, Mutex}, thread::JoinHandle, marker::PhantomData, cell::{Cell, UnsafeCell}, any::Any, panic::resume_unwind};
 
-use super::resolution_window::ResolutionWindow;
-use super::{registrar::Registrar, WitnessSource, WitnessSourceAwaitable};
-use crate::config::*;
-use crate::cs::traits::cs::{CSWitnessSource, DstBuffer};
-use crate::cs::{Place, Variable, VariableType};
-use crate::dag::awaiters::AwaitersBroker;
-use crate::dag::resolver_box::ResolverBox;
-use crate::dag::{awaiters, guide::*};
-use crate::field::SmallField;
-use crate::utils::{PipeOp, UnsafeCellEx};
-use itertools::Itertools;
-use std::ops::{Add, Sub, AddAssign};
-use std::panic::resume_unwind;
-use std::sync::atomic::{fence, AtomicBool, AtomicIsize};
-use std::sync::{Arc, Mutex};
-use std::thread::JoinHandle;
+use crate::{
+    field::SmallField,
+    dag::{
+        CircuitResolver,
+        WitnessSource,
+        WitnessSourceAwaitable,
+        awaiters::{self, AwaitersBroker}, resolver_box::ResolverBox, primitives::{Values, ExecOrder} },
+    config::CSResolverConfig,
+    cs::{
+        Place,
+        traits::cs::{ DstBuffer, CSWitnessSource }},
+    log,
+    utils::{ PipeOp as _, UnsafeCellEx }};
 
+use self::{resolution_window::ResolutionWindow, sorters::{ResolverSortingMode, ResolutionRecord}};
 
-pub(crate) type Mutarc<T> = Arc<Mutex<T>>;
+pub(crate) const PARANOIA: bool = false;
 
-
-
-
-
-#[derive(PartialEq, Eq, PartialOrd, Ord, Debug, Default, Clone, Copy)]
-pub struct OrderIx(u32);
-
-
-impl From<u32> for OrderIx {
-    fn from(value: u32) -> Self {
-        Self(value)
-    }
+/// Used to send notifications and data between the resolver, resolution window
+/// and the awaiters.
+pub struct ResolverComms {
+    pub exec_order_buffer_hint: AtomicIsize,
+    pub registration_complete: AtomicBool,
+    pub rw_panicked: AtomicBool,
+    pub rw_panic: Cell<Option<Box<dyn Any + Send + 'static>>>,
 }
 
-impl From<u64> for OrderIx {
-    fn from(value: u64) -> Self {
-        // This trait will not fail under normal circumstances.
-        debug_assert!(value < u32::MAX as u64);
-        Self(value as u32)
-    }
+#[derive(Debug)]
+struct Stats {
+    values_added: u64,
+    witnesses_added: u64,
+    registrations_added: u64,
+    started_at: std::time::Instant,
+    registration_time: std::time::Duration,
+    total_resolution_time: std::time::Duration,
 }
 
-impl From<usize> for OrderIx {
-    fn from(value: usize) -> Self {
-        // This trait will not fail under normal circumstances.
-        debug_assert!(value < u32::MAX as usize);
-        Self(value as u32)
+impl Stats {
+    fn new() -> Self {
+        Self {
+            values_added: 0,
+            witnesses_added: 0,
+            registrations_added: 0,
+            started_at: std::time::Instant::now(),
+            registration_time: std::time::Duration::from_secs(0),
+            total_resolution_time: std::time::Duration::from_secs(0),
+        }
     }
-}
-
-impl From<OrderIx> for u32 {
-    fn from(value: OrderIx) -> Self {
-        value.0
-    }
-}
-
-impl From<OrderIx> for u64 {
-    fn from(value: OrderIx) -> Self {
-        value.0 as u64
-    }
-}
-
-impl From<OrderIx> for usize {
-    fn from(value: OrderIx) -> Self {
-        value.0 as usize
-    }
-}
-
-impl From<OrderIx> for isize {
-    fn from(value: OrderIx) -> Self {
-        value.0 as isize
-    }
-}
-
-impl TrackId for OrderIx { }
-
-impl Add<i32> for OrderIx {
-    type Output = Self;
-
-    fn add(self, rhs: i32) -> Self::Output {
-        debug_assert!(rhs >= 0);
-        OrderIx(self.0 + rhs as u32)
-    }
-}
-
-impl AddAssign<i32> for OrderIx {
-    fn add_assign(&mut self, rhs: i32) {
-        *self = *self + rhs;
-    }
-}
-
-impl Add<u32> for OrderIx {
-    type Output = Self;
-
-    fn add(self, rhs: u32) -> Self::Output {
-        OrderIx(self.0 + rhs as u32)
-    }
-}
-
-impl AddAssign<u32> for OrderIx {
-        fn add_assign(&mut self, rhs: u32) {
-        *self = *self + rhs;
-    }
-}
-
-
-impl Add<usize> for OrderIx {
-    type Output = Self;
-
-    fn add(self, rhs: usize) -> Self::Output {
-        debug_assert!(rhs < u32::MAX as usize);
-        OrderIx(self.0 + rhs as u32)
-    }
-}
-
-impl Sub for OrderIx {
-    type Output = u32;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        self.0 - rhs.0
-    }
-}
-
-pub struct ExecOrder {
-    /// Represents the current size of the execution order. It is safe to execute the order up to
-    /// this value.
-    pub size: usize,
-    /// The order itself. The `len` of the vector behaves differently in record and playback mode.
-    /// Code agnostic to the mode can't rely on it.
-    pub items: Vec<OrderInfo<ResolverIx>>
 }
 
 /// Shared between the resolver, awaiters and the resolution window.
@@ -159,77 +71,316 @@ pub struct ResolverCommonData<V, T: Default> {
     pub awaiters_broker: AwaitersBroker<T>,
 }
 
+/// The data is tracked in the following manner:
+///
+/// `key ---> [values.variables/witnesses] ---> [resolvers_order] ---> [resolvers]`
+///
+/// Given a key, one can find a value and the metadata in `variables/witnesses`.
+/// The metadata contains the resolver order index which will produce a value for that item.
+/// The order index contains the index at which the resolver data is placed.
+///    Those indicies are not monotonic and act akin to pointers, and thus are
+///    Unsafe to work with.
 
+pub struct MtCircuitResolver<V: SmallField, RS: ResolverSortingMode<V>, CFG: CSResolverConfig> {
+    // registrar: Registrar,
 
-// region: ResolverIx
+    sorter: RS,
 
-#[derive(Copy, Clone, Default, PartialEq, Eq, PartialOrd, Ord, Debug)]
-pub struct ResolverIx(pub u32);
+    pub(crate) common: Arc<ResolverCommonData<V, RS::TrackId>>,
+    // pub(crate) options: CircuitResolverOpts,
+    
+    comms: Arc<ResolverComms>,
+    // pub(crate) guide: BufferGuide<ResolverIx, Cfg>,
 
-pub enum ResolverIxType {
-    Jump,
-    Resolver,
+    resolution_window_handle: Option<JoinHandle<()>>,
+
+    stats: Stats,
+    call_count: u32,
+    debug_track: Vec<Place>,
+    phantom: PhantomData<CFG>,
 }
 
-impl ResolverIx {
-    // Resolver box uses `sizeof` to determine the size of the allocations,
-    // and operates on pointers or _size primitives, which always have lsb == 0
-    // in their sizes, thus we can use the lsb to store type info.
-    const TYPE_MASK: u32 = 1;
+unsafe impl<V, RS, CFG> Send for MtCircuitResolver<V, RS, CFG> 
+where 
+    V: SmallField,
+    RS: ResolverSortingMode<V>,
+    CFG: CSResolverConfig {}
 
-    pub fn get_type(self) -> ResolverIxType {
-        match self.0 & Self::TYPE_MASK == 0 {
-            true => ResolverIxType::Resolver,
-            false => ResolverIxType::Jump,
+unsafe impl<V, RS, CFG> Sync for MtCircuitResolver<V, RS, CFG> 
+where 
+    V: SmallField,
+    RS: ResolverSortingMode<V>,
+    CFG: CSResolverConfig {}
+
+impl< F, RS, CFG> CircuitResolver<F, CFG> for MtCircuitResolver<F, RS, CFG> 
+where 
+    F: SmallField,
+    RS: ResolverSortingMode<F> + 'static,
+    CFG: CSResolverConfig 
+{
+    type Arg = RS::Arg;
+
+    fn new(args: Self::Arg) -> Self {
+        Self::new(args)
+    }
+
+    fn set_value(&mut self, key: Place, value: F) {
+        self.set_value(key, value)
+    }
+
+    fn add_resolution<Fn>(&mut self, inputs: &[Place], outputs: &[Place], f: Fn)
+    where
+        Fn: FnOnce(&[F], &mut DstBuffer<'_, '_, F>) + Send + Sync {
+        self.add_resolution(inputs, outputs, f)
+    }
+
+    fn wait_till_resolved(&mut self) {
+        self.wait_till_resolved()
+    }
+
+    fn clear(&mut self) {
+        self.clear()
+    }
+}
+
+impl<V: SmallField, RS: ResolverSortingMode<V>, CFG: CSResolverConfig> MtCircuitResolver<V, RS, CFG> {
+    pub fn new(opts: RS::Arg) -> Self {
+        let threads =
+                std::env::var("BOOJUM_CR_THREADS")
+                .map_err(|_| "")
+                .and_then(|x| x.parse().map_err(|_| ""))
+                .unwrap_or(3);
+
+        let debug_track = vec![];
+
+        if cfg!(cr_paranoia_mode) || PARANOIA {
+            log!("Contains tracked keys {:?} ", debug_track);
+        }
+
+        let comms = ResolverComms {
+            exec_order_buffer_hint: AtomicIsize::new(0),
+            registration_complete: AtomicBool::new(false),
+            rw_panicked: AtomicBool::new(false),
+            rw_panic: Cell::new(None),
+        }.to(Arc::new);
+
+        let (sorter, common) = RS::new(opts, comms.clone(), &debug_track);
+
+        Self {
+            // options: opts,
+            call_count: 0,
+            sorter,
+            // guide: BufferGuide::new(opts.desired_parallelism),
+            // registrar: Registrar::new(),
+            comms: comms.clone(),
+
+            resolution_window_handle: ResolutionWindow::<V, RS::TrackId, RS::Config>::run(
+                comms,
+                common.clone(),
+                &debug_track,
+                threads,
+            )
+            .to(Some),
+
+            common,
+            stats: Stats::new(),
+            debug_track,
+            phantom: PhantomData,
         }
     }
 
-    fn new_jump(value: u32) -> Self {
-        Self(value | Self::TYPE_MASK)
+    pub fn set_value(&mut self, key: Place, value: V) {
+        self.sorter.set_value(key, value)
     }
 
-    pub fn new_resolver(value: u32) -> Self {
-        Self(value)
+    pub fn add_resolution<F>(&mut self, inputs: &[Place], outputs: &[Place], f: F)
+    where
+        F: FnOnce(&[V], &mut DstBuffer<'_, '_, V>) + Send + Sync,
+    {
+        self.sorter.add_resolution(inputs, outputs, f)
     }
 
-    pub fn normalized(&self) -> usize {
-        (!Self::TYPE_MASK & self.0) as usize
+    pub fn wait_till_resolved(&mut self) {
+        self.wait_till_resolved_impl(true);
+    }
+
+    pub fn wait_till_resolved_impl(&mut self, report: bool) {
+        if self
+            .comms
+            .registration_complete
+            .load(std::sync::atomic::Ordering::Relaxed)
+        {
+            return;
+        }
+
+        self.sorter.final_flush();
+
+        self.stats.registration_time = self.stats.started_at.elapsed();
+
+        self
+            .comms
+            .registration_complete
+            .store(true, std::sync::atomic::Ordering::Relaxed);
+
+        self.resolution_window_handle
+            .take()
+            .expect("Attempting to join resolution window handler for second time.")
+            .join()
+            .unwrap(); // Just propagate panics. Those are unhandled, unlike the ones from `rw_panic`.
+
+        self.stats.total_resolution_time = self.stats.started_at.elapsed();
+
+        // Propage panic from the resolution window handler.
+        if self
+            .comms
+            .rw_panicked
+            .load(std::sync::atomic::Ordering::Relaxed)
+        {
+            if let Some(e) = self.comms.rw_panic.take() {
+                resume_unwind(e);
+            } else {
+                log!("Resolution window panicked, but no panic payload stored.");
+                return;
+            }
+        }
+
+        match report {
+            true => {
+                log!("CR stats {:#?}", self.stats);
+            }
+            false if cfg!(test) || cfg!(debug_assertions) => {
+                print!(" resolution time {:?}...", self.stats.total_resolution_time);
+            }
+            _ => {}
+        }
+        
+        self.sorter.write_sequence();
+
+        if cfg!(cr_paranoia_mode) || PARANOIA {
+            log!("CR {:?}", unsafe {
+                self.common.awaiters_broker.stats.u_deref()
+            });
+        }
+    }
+
+    pub fn retrieve_sequence(&mut self) -> &ResolutionRecord {
+        assert!(self.comms.registration_complete.load(std::sync::atomic::Ordering::Relaxed));
+        self.sorter.retrieve_sequence()
+    }
+
+    pub fn clear(&mut self) {
+        // TODO: implement
     }
 }
 
-impl AddAssign for ResolverIx {
-    fn add_assign(&mut self, rhs: Self) {
-        self.0 = rhs.0;
+impl<V: SmallField, RS: ResolverSortingMode<V> + 'static, CFG: CSResolverConfig> 
+    WitnessSource<V> for MtCircuitResolver<V, RS, CFG> {
+    const PRODUCES_VALUES: bool = true;
+
+    fn try_get_value(&self, variable: Place) -> Option<V> {
+
+        let (v, md) = unsafe { self.common.values.u_deref().get_item_ref(variable) };
+
+        match md.is_resolved() {
+            true => {
+                fence(std::sync::atomic::Ordering::Acquire);
+                Some(*v)
+            }
+            false => None,
+        }
+    }
+
+    fn get_value_unchecked(&self, variable: Place) -> V {
+        // TODO: Should this fn be marked as unsafe?
+
+        // Safety: Dereferencing as & in &self context.
+        let (r, md) = unsafe { self.common.values.u_deref().get_item_ref(variable) };
+        // log!("gvu: {:0>8} -> {}", variable.0, r);
+
+        debug_assert!(
+            md.is_resolved(),
+            "Attempted to get value of unresolved variable."
+        );
+
+        *r
     }
 }
 
-impl Sub for ResolverIx {
-    type Output = u32;
+impl<V: SmallField, RS: ResolverSortingMode<V> + 'static, CFG: CSResolverConfig> 
+    CSWitnessSource<V> for MtCircuitResolver<V, RS, CFG> {}
 
-    fn sub(self, origin: Self) -> Self::Output {
-        self.0 - origin.0
+impl<V: SmallField, RS: ResolverSortingMode<V> + 'static, CFG: CSResolverConfig> 
+    WitnessSourceAwaitable<V> for MtCircuitResolver<V, RS, CFG> {
+    type Awaiter<'a> = awaiters::Awaiter<'a, RS::TrackId>;
+
+    fn get_awaiter<const N: usize>(&mut self, vars: [Place; N]) -> awaiters::Awaiter<RS::TrackId> {
+        // Safety: We're only getting the metadata address for an item, which is
+        // immutable and the max_tracked value, which isn't but read only once
+        // for the duration of the reference.
+        let values = unsafe { self.common.values.u_deref() };
+
+        if values.max_tracked < vars.iter().map(|x| x.as_any_index()).max().unwrap() as i64 {
+            panic!("The awaiter will never resolve since the awaited variable can't be computed based on currently available registrations. You have holes!!!");
+        }
+
+        // We're picking the item that will be resolved last among other inputs.
+        let md = vars
+            .into_iter()
+            .map(|x| &values.get_item_ref(x).1)
+            .max_by_key(|x| x.tracker)
+            .unwrap();
+
+        let r = awaiters::AwaitersBroker::register(
+            &self.common.awaiters_broker,
+            &self.comms,
+            md,
+        );
+
+        self.sorter.flush();
+
+        r
     }
 }
 
-impl AddAssign<u32> for ResolverIx {
-    fn add_assign(&mut self, rhs: u32) {
-        self.0 = rhs;
+// impl Drop for CircuitResolver
+
+impl<V: SmallField, RS: ResolverSortingMode<V>, CFG: CSResolverConfig> 
+    Drop for MtCircuitResolver<V, RS, CFG> {
+    fn drop(&mut self) {
+        if cfg!(test) || cfg!(debug_assertions) {
+            print!("Starting drop of CircuitResolver (If this hangs, it's bad)...");
+        }
+        self.wait_till_resolved_impl(false);
+
+        if cfg!(test) || cfg!(debug_assertions) {
+            log!("ok");
+        }
     }
 }
 
-// endregion
-
-// region: Values
-
-
-
-// endregion
 
 #[cfg(test)]
 mod test {
-    use crate::{dag::{Awaiter, sorter_runtime::RuntimeResolverSorter, sorter_playback::PlaybackResolverSorter, ResolutionRecordStorage, CircuitResolverOpts, TestRecordStorage, resolvers::MtCircuitResolver}, utils::PipeOp};
-    use std::{collections::VecDeque, hint::spin_loop, time::Duration, rc::Rc};
+    use std::collections::VecDeque;
+    use std::hint::spin_loop;
+    use std::rc::Rc;
+    use std::sync::Mutex;
 
+    use itertools::Itertools as _;
+
+
+    use crate::config::Resolver;
+    use crate::cs::Place;
+    use crate::cs::traits::cs::DstBuffer;
+    use crate::dag::resolvers::mt::sorters::sorter_playback::PlaybackResolverSorter;
+    use crate::dag::{CircuitResolverOpts, WitnessSource as _, WitnessSourceAwaitable as _, Awaiter};
+    use crate::dag::resolvers::MtCircuitResolver;
+    use crate::dag::resolvers::mt::sorters::ResolverSortingMode;
+    use crate::dag::resolvers::mt::sorters::sorter_runtime::*;
+
+    use crate::field::SmallField;
+    use crate::log;
+    use crate::utils::PipeOp as _;
     use crate::{
         config::DoPerformRuntimeAsserts,
         cs::Variable,
@@ -237,10 +388,25 @@ mod test {
     };
 
     use super::*;
-    use super::super::sorter_runtime::*;
+    use super::sorters::{ResolutionRecordWriter, ResolutionRecordSource};
 
     type F = GoldilocksField;
-    type CFG = Resolver<DoPerformRuntimeAsserts>;
+    type Cfg = Resolver<DoPerformRuntimeAsserts>;
+
+    pub struct TestRecordStorage {
+        record: std::rc::Rc<ResolutionRecord>
+    }
+
+    impl ResolutionRecordWriter for TestRecordStorage {
+        fn store(&mut self, _record: &ResolutionRecord) {
+        }
+    }
+
+    impl ResolutionRecordSource for TestRecordStorage {
+        fn get(&self) -> &ResolutionRecord {
+            &self.record
+        }
+    }
 
     #[test]
     fn playground() {
@@ -257,7 +423,7 @@ mod test {
     }
 
     fn tracks_values_populate<F: SmallField, RS: ResolverSortingMode<F>>(
-        resolver: &mut MtCircuitResolver<F, RS, CFG>, limit: u64)
+        resolver: &mut MtCircuitResolver<F, RS, Cfg>, limit: u64)
     {
         for i in 0..limit {
             let a = Place::from_variable(Variable::from_variable_index(i));
@@ -270,7 +436,7 @@ mod test {
     fn tracks_values_record_mode() {
         let limit = 10;
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(
                 CircuitResolverOpts {
                     max_variables: 10,
                     desired_parallelism: 16,
@@ -292,7 +458,7 @@ mod test {
     fn tracks_values_playback_mode() {
         let limit = 10;
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(
                 CircuitResolverOpts {
                     max_variables: 10,
                     desired_parallelism: 16,
@@ -306,7 +472,9 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                // ActiveRecordingResolverSorter<F, CFG, TestRecordStorage>,
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>,
+                Cfg>
             ::new(rs);
 
         tracks_values_populate(&mut storage, limit);
@@ -320,7 +488,7 @@ mod test {
     }
 
     fn resolves_populate<F: SmallField, RS: ResolverSortingMode<F>>(
-        resolver: &mut MtCircuitResolver<F, RS, CFG>) -> (Place, Place)
+        resolver: &mut MtCircuitResolver<F, RS, Cfg>) -> (Place, Place)
     {
 
         let res_fn = |ins: &[F], outs: &mut DstBuffer<F>| {
@@ -340,7 +508,7 @@ mod test {
     #[test]
     fn resolves_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -358,7 +526,7 @@ mod test {
     #[test]
     fn resolves_playback_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -374,7 +542,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         let (init_var, dep_var) = resolves_populate(&mut storage);
@@ -388,7 +556,7 @@ mod test {
     }
 
     fn resolves_siblings_populate<F: SmallField, RS: ResolverSortingMode<F>>(
-        resolver: &mut MtCircuitResolver<F, RS, CFG>) -> ((Place, Place), (Place, Place))
+        resolver: &mut MtCircuitResolver<F, RS, Cfg>) -> ((Place, Place), (Place, Place))
     {
         let res_fn = |ins: &[F], outs: &mut DstBuffer<F>| {
             let mut x = ins[0];
@@ -413,7 +581,7 @@ mod test {
     #[test]
     fn resolves_siblings_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -436,7 +604,7 @@ mod test {
     #[test]
     fn resolves_siblings_playback_mode() {
         let mut storage =
-        MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+        MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
             max_variables: 100,
             desired_parallelism: 16,
         });
@@ -450,7 +618,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         let ((init_var1, dep_var1), (init_var2, dep_var2)) 
@@ -469,7 +637,7 @@ mod test {
     }
 
     fn resolves_descendants_populate<F: SmallField, RS: ResolverSortingMode<F>>(
-        resolver: &mut MtCircuitResolver<F, RS, CFG>) -> Place
+        resolver: &mut MtCircuitResolver<F, RS, Cfg>) -> Place
     {
         let res_fn = |ins: &[F], outs: &mut DstBuffer<F>| {
             let mut x = ins[0];
@@ -494,7 +662,7 @@ mod test {
     #[test]
     fn resolves_descendants_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 2,
             });
@@ -513,7 +681,7 @@ mod test {
     // #[ignore = "temp"]
     fn resolves_descendants_playback_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 2,
             });
@@ -527,7 +695,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         let dep_var3 = resolves_descendants_populate(&mut storage);
@@ -543,7 +711,7 @@ mod test {
     #[test]
     fn resolves_with_context() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -578,7 +746,7 @@ mod test {
     #[test]
     fn resolves_and_drops_context_after() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -589,7 +757,7 @@ mod test {
         storage.set_value(init_var, F::from_u64_with_reduction(123));
 
         struct DroppedContext {
-            times_invoked: Mutarc<u32>,
+            times_invoked: Arc<Mutex<u32>>,
             value: u64,
         }
 
@@ -627,7 +795,7 @@ mod test {
     fn awaiter_returns_for_resolved_value_record_mode() {
         let limit = 1 << 13;
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit * 5,
                 desired_parallelism: 2048,
             });
@@ -666,7 +834,7 @@ mod test {
     fn awaiter_returns_for_resolved_value_playback_mode_impl(limit: usize, desired_parallelism: u32) {
         let limit = 1 << limit;
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit * 5,
                 desired_parallelism,
             });
@@ -680,7 +848,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         populate(&mut storage, limit);
@@ -706,7 +874,7 @@ mod test {
     #[test]
     fn awaiter_returns_after_finish_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -735,7 +903,7 @@ mod test {
     #[test]
     fn awaiter_returns_after_finish_playback_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -760,7 +928,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(init_var, F::from_u64_with_reduction(123));
@@ -781,7 +949,7 @@ mod test {
     #[test]
     fn awaiter_returns_for_unexpropriated() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -807,7 +975,7 @@ mod test {
     #[test]
     fn awaiter_blocks_before_resolved() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -815,7 +983,7 @@ mod test {
         let mut notch = std::time::Instant::now();
 
         let res_fn = |ins: &[F], outs: &mut DstBuffer<F>| {
-            std::thread::sleep(Duration::from_secs(1));
+            std::thread::sleep(std::time::Duration::from_secs(1));
             notch = std::time::Instant::now();
             outs.push(ins[0]);
         };
@@ -837,7 +1005,7 @@ mod test {
     #[test]
     fn resolution_after_awaiter_is_supported_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -876,7 +1044,7 @@ mod test {
         };
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
         });
@@ -894,7 +1062,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(init_var, F::from_u64_with_reduction(123));
@@ -915,7 +1083,7 @@ mod test {
     #[test]
     fn try_get_value_returns_none_before_resolve_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -947,7 +1115,7 @@ mod test {
         let dep_var = Place::from_variable(Variable::from_variable_index(1));
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -962,7 +1130,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(init_var, F::from_u64_with_reduction(123));
@@ -975,7 +1143,7 @@ mod test {
     #[test]
     fn try_get_value_returns_some_after_resolve_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1008,7 +1176,7 @@ mod test {
         let dep_var = Place::from_variable(Variable::from_variable_index(1));
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
         });
@@ -1022,7 +1190,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(init_var, F::from_u64_with_reduction(123));
@@ -1037,7 +1205,7 @@ mod test {
     #[test]
     fn try_get_value_returns_some_after_wait_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1063,7 +1231,7 @@ mod test {
     #[test]
     fn try_get_value_returns_some_after_wait_playback_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1087,7 +1255,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(init_var, F::from_u64_with_reduction(123));
@@ -1103,7 +1271,7 @@ mod test {
     #[test]
     fn try_get_value_returns_none_on_untracked() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1130,7 +1298,7 @@ mod test {
     #[should_panic]
     fn panic_in_resolution_function_is_propagated_through_cr_waiting() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1155,7 +1323,7 @@ mod test {
     #[should_panic]
     fn panic_in_resolution_function_is_propagated_through_awaiter() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1177,7 +1345,7 @@ mod test {
     #[test]
     fn non_chronological_resolution_record_mode() {
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1204,8 +1372,6 @@ mod test {
         storage.wait_till_resolved();
 
         let result = storage.try_get_value(var_5);
-
-        let record = storage.retrieve_sequence();
 
         assert_eq!(Some(F::from_u64_with_reduction(105)), result);
     }
@@ -1227,7 +1393,7 @@ mod test {
         let var_5 = Place::from_variable(Variable::from_variable_index(4));
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: 100,
                 desired_parallelism: 16,
             });
@@ -1245,7 +1411,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         storage.set_value(var_4, F::from_u64_with_reduction(7));
@@ -1262,7 +1428,7 @@ mod test {
     }
 
     fn correctness_simple_linear_populate<F: SmallField, RS: ResolverSortingMode<F>>(
-        resolver: &mut MtCircuitResolver<F, RS, CFG>, limit: usize)
+        resolver: &mut MtCircuitResolver<F, RS, Cfg>, limit: usize)
     {
         let mut var_idx = 0;
 
@@ -1305,7 +1471,7 @@ mod test {
         let limit = 1 << 10;
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit * 5,
                 desired_parallelism: 32,
             });
@@ -1350,7 +1516,7 @@ mod test {
         let limit = 1 << 10;
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit * 5,
                 desired_parallelism: 32,
             });
@@ -1363,7 +1529,7 @@ mod test {
         let mut storage =
             MtCircuitResolver::<
                 F, 
-                PlaybackResolverSorter<F, TestRecordStorage, CFG>, CFG>
+                PlaybackResolverSorter<F, TestRecordStorage, Cfg>, Cfg>
             ::new(rs);
 
         correctness_simple_linear_populate(&mut storage, limit);
@@ -1402,7 +1568,7 @@ mod test {
     }
 
 
-    fn populate<RS: ResolverSortingMode<F>>(storage: &mut MtCircuitResolver<F, RS, CFG>, limit: usize) {
+    fn populate<RS: ResolverSortingMode<F>>(storage: &mut MtCircuitResolver<F, RS, Cfg>, limit: usize) {
 
         let mut var_idx = 0u64;
         for _ in 0..limit {
@@ -1456,12 +1622,11 @@ mod benches {
 
     use super::*;
     use crate::{
-        cs::Variable,
-        dag::{Awaiter, sorter_runtime::RuntimeResolverSorter, CircuitResolverOpts, resolvers::MtCircuitResolver},
-        field::{goldilocks::GoldilocksField, Field},
+        cs::{Variable, Place, traits::cs::DstBuffer},
+        field::{goldilocks::GoldilocksField, Field}, config::{Resolver, DoPerformRuntimeAsserts}, log, dag::{resolvers::mt::sorters::sorter_runtime::RuntimeResolverSorter, CircuitResolverOpts, Awaiter as _},
     };
     type F = GoldilocksField;
-    type CFG = Resolver<DoPerformRuntimeAsserts>;
+    type Cfg = Resolver<DoPerformRuntimeAsserts>;
 
     #[test]
     #[ignore = ""]
@@ -1482,7 +1647,7 @@ mod benches {
     fn synth_bench_1() {
         let limit = 1 << 25;
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit * 5,
                 desired_parallelism: 2048,
             });
@@ -1555,7 +1720,7 @@ mod benches {
         let limit = 1 << 4;
 
         let mut storage =
-            MtCircuitResolver::<F, RuntimeResolverSorter<F, CFG>, CFG>::new(CircuitResolverOpts {
+            MtCircuitResolver::<F, RuntimeResolverSorter<F, Cfg>, Cfg>::new(CircuitResolverOpts {
                 max_variables: limit + 1,
                 desired_parallelism: 16,
             });
