+++ src/dag/resolvers/null.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use crate::{dag::{CircuitResolver, WitnessSource, WitnessSourceAwaitable, awaiters::Awaiter, resolver::OrderIx}, field::SmallField, config::CSResolverConfig, cs::traits::cs::CSWitnessSource};
+use crate::{dag::{CircuitResolver, WitnessSource, WitnessSourceAwaitable, awaiters::Awaiter, primitives::OrderIx}, field::SmallField, config::CSResolverConfig, cs::traits::cs::CSWitnessSource};
 
 pub struct NullCircuitResolver<F, CFG> {
     phantom: PhantomData<(F, CFG)>
@@ -10,11 +10,11 @@ pub struct NullCircuitResolver<F, CFG> {
 impl<F: SmallField, CFG: CSResolverConfig> WitnessSource<F> for NullCircuitResolver<F, CFG> {
     const PRODUCES_VALUES: bool = false;
 
-    fn try_get_value(&self, variable: crate::cs::Place) -> Option<F> {
+    fn try_get_value(&self, _variable: crate::cs::Place) -> Option<F> {
         panic!("Null resolver.");
     }
 
-    fn get_value_unchecked(&self, variable: crate::cs::Place) -> F {
+    fn get_value_unchecked(&self, _variable: crate::cs::Place) -> F {
         panic!("Null resolver.");
     }
 }
@@ -22,7 +22,7 @@ impl<F: SmallField, CFG: CSResolverConfig> WitnessSource<F> for NullCircuitResol
 impl<F: SmallField, CFG: CSResolverConfig> WitnessSourceAwaitable<F> for NullCircuitResolver<F, CFG> {
     type Awaiter<'a> = Awaiter<'a, OrderIx>;
 
-    fn get_awaiter<const N: usize>(&mut self, vars: [crate::cs::Place; N]) -> Self::Awaiter<'_> {
+    fn get_awaiter<const N: usize>(&mut self, _vars: [crate::cs::Place; N]) -> Self::Awaiter<'_> {
         panic!("Null resolver.");
     }
 }
@@ -38,15 +38,15 @@ impl<
 {
     type Arg = ();
 
-    fn new(args: Self::Arg) -> Self {
+    fn new(_args: Self::Arg) -> Self {
         panic!("Null resolver.");
     }
 
-    fn set_value(&mut self, key: crate::cs::Place, value: F) {
+    fn set_value(&mut self, _key: crate::cs::Place, _value: F) {
         panic!("Null resolver.");
     }
 
-    fn add_resolution<Fn>(&mut self, inputs: &[crate::cs::Place], outputs: &[crate::cs::Place], f: Fn)
+    fn add_resolution<Fn>(&mut self, _inputs: &[crate::cs::Place], _outputs: &[crate::cs::Place], _f: Fn)
         where
         Fn: FnOnce(&[F], &mut crate::cs::traits::cs::DstBuffer<'_, '_, F>) + Send + Sync {
         panic!("Null resolver");
