+++ src/gadgets/poseidon2/mod.rs
@@ -456,7 +456,6 @@ mod test {
     use crate::config::CSConfig;
     use crate::cs::gates::*;
     use crate::dag::CircuitResolverOpts;
-    use crate::dag::sorter_runtime::RuntimeResolverSorter;
     use crate::log;
     use crate::worker::Worker;
